test: share the fake-clock preload across packages
Two packages install a fake-clock Deno `--preload` that replaces `Deno.test`
so tests run under controllable time: the runner suite and the html reconciler
suite. Each carried its own copy of the timer-faking machinery — the caller
classification that sorts a positive-delay `setTimeout` by whether a `src/` or a
`test/` stack frame scheduled it, the zero-delay kick that drains scheduler
dispatch on a real macrotask, the drain-to-fixpoint `settle`, the auto-advance
pump, `tick`/`reset`, and the runaway cap. The runner copy is about three
hundred lines. More packages (ui, background-piece-service, utils) are being
converted to the same approach, each about to add another copy on top.

Factor that machinery into one shared harness that packages configure instead
of duplicate. `@commonfabric/test-support/clock-preload` exports
`installFakeClock({ mode, realClockFiles })`. Two modes cover what exists today:

- "auto-advance" (runner): a positive-delay timer armed from production code
  advances logical time on its own so the runtime's reactive waits resolve
  without real sleeping, while a wall-clock sleep armed from a test file freezes
  and its waiter deadlocks. It also fakes `setInterval`, `Date.now`, and
  `performance.now`, exposes a global `clock` with `settle`/`tick`/`reset`, and
  drains once after the test body.
- "freeze-all" (html): every positive-delay timer freezes regardless of caller,
  only `setTimeout`/`clearTimeout` are replaced, and the control is `t.settle`.

Each package's `test/clock-preload.ts` becomes a one-call wrapper that selects a
mode and, for the runner, passes its own list of test files kept on the real
clock (with the reasons they cannot run under the fake one). The harness reads
its own basename from `import.meta.url` for the stack-frame caller
classification, so the wrappers — which share that basename — and the shared
file are both skipped when deciding who scheduled a timer.

The runner suite (1068 tests) and the html suite (43 tests) both stay green
under the shared harness with repo-wide fmt/lint clean. The ambient type
declarations (`test/clock.d.ts` in each package) stay put, since they augment
the global scope that each package's own `deno check` type-checks.
